.history/apis/requirements_20251204171052.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
scikit-learn
numpy
apscheduler
python-dotenv

